+++ package.json
@@ -52,6 +52,7 @@
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
     "redux-saga": "^1.1.3",
+    "reselect": "^4.0.0",
     "yup": "^0.29.3"
   },
   "scripts": {
